% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_admin_names.R
\name{apply_case_mapping}
\alias{apply_case_mapping}
\title{Apply Case Mapping to Data Frame}
\usage{
apply_case_mapping(df, case_mapping, levels)
}
\arguments{
\item{df}{Data frame to apply case mapping to}

\item{case_mapping}{List of case mappings for each level}

\item{levels}{Vector of level column names}
}
\description{
Internal helper function to apply original case from lookup data
to the matched values in the target data frame.
}
\keyword{internal}
