% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_rate.R
\name{calculate_plot_dimensions}
\alias{calculate_plot_dimensions}
\title{Calculate appropriate plot dimensions based on data}
\usage{
calculate_plot_dimensions(plot_data, x_var, y_var = NULL)
}
\arguments{
\item{plot_data}{Data frame with the plotted data}

\item{x_var}{X-axis variable}

\item{y_var}{Y-axis variable (can be NULL)}
}
\value{
List with width and height values
}
\description{
Calculate appropriate plot dimensions based on data
}
