% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dhs_indicators.R
\name{check_dhs_indicators}
\alias{check_dhs_indicators}
\title{Check DHS Indicator List from API}
\usage{
check_dhs_indicators(
  countryIds = NULL,
  indicatorIds = NULL,
  surveyIds = NULL,
  surveyYear = NULL,
  surveyYearStart = NULL,
  surveyYearEnd = NULL,
  surveyType = NULL,
  surveyCharacteristicIds = NULL,
  tagIds = NULL,
  returnFields = c("IndicatorId", "Label", "Definition", "MeasurementType"),
  perPage = NULL,
  page = NULL,
  f = "json"
)
}
\arguments{
\item{countryIds}{DHS country code(s), e.g., "EG"}

\item{indicatorIds}{Specific indicator ID(s)}

\item{surveyIds}{Survey ID(s)}

\item{surveyYear}{Exact year}

\item{surveyYearStart}{Start of year range}

\item{surveyYearEnd}{End of year range}

\item{surveyType}{DHS survey type (e.g., "DHS", "MIS")}

\item{surveyCharacteristicIds}{Filter by survey characteristic ID}

\item{tagIds}{Filter by tag ID}

\item{returnFields}{Fields to return (default: IndicatorId, Label, Definition)}

\item{perPage}{Max results per page (default = 500)}

\item{page}{Specific page to return (default = 1)}

\item{f}{Format (default = "json")}
}
\value{
A data.frame of indicators
}
\description{
Check DHS Indicator List from API
}
