% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rasters.R
\name{clean_filenames}
\alias{clean_filenames}
\title{Clean Filenames}
\usage{
clean_filenames(filenames)
}
\arguments{
\item{filenames}{A character vector of full file paths or basenames.}
}
\value{
A character vector of cleaned filenames. If full paths are
provided, the function returns the cleaned filenames with the full path
intact.
}
\description{
This function cleans a vector of filenames by removing common standalone
numbers, condensing multiple separators, and ensuring clean formats.
}
