% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_rate_plot.R
\name{create_common_elements}
\alias{create_common_elements}
\title{Create common ggplot elements}
\usage{
create_common_elements(fill_var, fill_limits, use_reprate = TRUE)
}
\arguments{
\item{fill_var}{Fill variable name}

\item{fill_limits}{Limits for the fill scale}

\item{use_reprate}{Whether to use reporting rate colors}
}
\value{
List of ggplot elements to apply to plots
}
\description{
Create common ggplot elements
}
