% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_era5.R
\name{get_era5_metadata}
\alias{get_era5_metadata}
\title{Get Metadata from ERA5 NetCDF File}
\usage{
get_era5_metadata(nc_file)
}
\arguments{
\item{nc_file}{Character. Path to ERA5 NetCDF file.}
}
\value{
A list containing:
\describe{
\item{resolution}{Spatial resolution in degrees (lon, lat)}
\item{extent}{Bounding box (lon_min, lon_max, lat_min, lat_max)}
\item{dimensions}{List of dimension information}
\item{variables}{Tibble of variable metadata (name,
long_name, units, dimensions)}
\item{global_attrs}{List of global attributes}
}
}
\description{
Extracts comprehensive metadata from an ERA5 NetCDF file including
resolution, variables, units, dimensions, and global attributes.
}
\examples{
\dontrun{
file <- "era5_data/monthly_single_levels_era5_202001.nc"
metadata <- get_era5_metadata(file)
print(metadata$resolution)
print(metadata$variables)
}

}
