% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_rate_plot.R
\name{get_translated_terms}
\alias{get_translated_terms}
\title{Get translated terms for plot filenames}
\usage{
get_translated_terms(language, x_var, vars_of_interest, data)
}
\arguments{
\item{language}{Language code}

\item{x_var}{X-axis variable}

\item{vars_of_interest}{Variables being visualized}

\item{data}{Original data for year range}
}
\value{
List of translated terms
}
\description{
Get translated terms for plot filenames
}
