% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_admin_names.R
\name{get_user_identity}
\alias{get_user_identity}
\title{Get User Identity}
\usage{
get_user_identity()
}
\value{
Character string with the user identity
}
\description{
Helper function to get the user identity from various environment variables
}
\keyword{internal}
