% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_bin.R
\name{list_palettes}
\alias{list_palettes}
\title{List available palette names}
\usage{
list_palettes()
}
\value{
Character vector of available palette names.
}
\description{
List available palette names
}
\examples{
list_palettes()

}
