% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean2}
\alias{mean2}
\title{Calculate mean with automatic NA handling}
\usage{
mean2(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The mean of all non-NA values in x
}
\description{
A wrapper around base::mean() that automatically removes NA
values
}
\examples{
mean2(c(1, 2, NA, 4)) # Returns 2.333333
mean2(c(NA, NA, 10, 20)) # Returns 15
}
