% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{median2}
\alias{median2}
\title{Calculate median with automatic NA handling}
\usage{
median2(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The median of all non-NA values in x
}
\description{
A wrapper around base::median() that automatically removes NA
values
}
\examples{
median2(c(1, 2, NA, 4)) # Returns 2.333333
median2(c(NA, NA, 10, 20)) # Returns 15
}
