% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_plot.R
\name{missing_plot}
\alias{missing_plot}
\title{Plot Missing data over time}
\usage{
missing_plot(
  data,
  x_var,
  y_var = NULL,
  miss_vars = NULL,
  use_rep_rate = FALSE,
  save_plot = FALSE,
  plot_path = NULL,
  plot_name = "missing_plot.png",
  plot_width = 10,
  plot_height = 8,
  plot_dpi = 300
)
}
\arguments{
\item{data}{A data frame containing the data to be visualized. Must include
columns specified in 'x_var', 'y_var', and 'vars'.}

\item{x_var}{A character string specifying the time variable in 'data'
(e.g., "year", "month"). Must be provided.}

\item{y_var}{An optional character string specifying the grouping
variable in 'data' (e.g., "state"). If provided, only one variable can be
specified in 'vars'.}

\item{miss_vars}{An optional character vector specifying the variables
to be visualized in 'data'. If NULL, all variables except 'x_var' and
'y_var' will be used.}

\item{use_rep_rate}{A logical value. If TRUE, the reporting rate is
visualized; otherwise, the proportion of missing data is visualized.
Defaults to FALSE}

\item{save_plot}{A logical value. If TRUE, the plot will be saved to the
specified path. Defaults to FALSE.}

\item{plot_path}{A character string specifying the path where the plot should
be saved. Required if save_plot is TRUE.}

\item{plot_name}{A character string specifying the name of the plot file.
Defaults to "missing_plot.png".}

\item{plot_width}{Numeric value for the width of the saved plot in inches.
Defaults to 10.}

\item{plot_height}{Numeric value for the height of the saved plot in inches.
Defaults to 8.}

\item{plot_dpi}{Numeric value for the resolution of the saved plot.
Defaults to 300.}
}
\value{
A ggplot2 object representing the tile plot.
}
\description{
This function visualizes the proportion of missing data or reporting rate for
specified variables in a dataset. It creates a tile plot where the x-axis can
represent any categorical time such as time (e.g., year, month), and the
y-axis can represents either variables or groupings (e.g., state). The
output can further be manipulated to one's needs.
}
\examples{


# get path
path <- system.file(
        "extdata",
        "fake_epi_df_togo.rds",
         package = "epiCleanr")

# get example data
fake_epi_df_togo <- read(path)

# Check missing data by year
result <- missing_plot(fake_epi_df_togo,
             x_var = "year", use_rep_rate = FALSE)

# Save the plot
result <- missing_plot(fake_epi_df_togo,
             x_var = "year",
             save_plot = TRUE,
             plot_path = tempdir(),
             plot_name = "missing_data_plot.png")
}
