% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum2}
\alias{sum2}
\title{Sum values with automatic NA handling}
\usage{
sum2(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The sum of all non-NA values in x
}
\description{
A wrapper around base::sum() that automatically removes NA
values
}
\examples{
sum2(c(1, 2, NA, 4)) # Returns 7
sum2(c(NA, NA, 5, 10)) # Returns 15
}
