% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_rate_plot.R
\name{translate_plot_labels}
\alias{translate_plot_labels}
\title{Translate plot labels to specified language}
\usage{
translate_plot_labels(plot, language)
}
\arguments{
\item{plot}{A ggplot2 object}

\item{language}{Target language code}
}
\value{
The plot with translated labels
}
\description{
Translate plot labels to specified language
}
