% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vdigest}
\alias{vdigest}
\title{Vectorized version of digest::digest}
\usage{
vdigest(
  object,
  algo = c("md5", "sha1", "crc32", "sha256", "sha512", "xxhash32", "xxhash64",
    "murmur32", "spookyhash", "blake3", "crc32c", "xxh3_64", "xxh3_128"),
  serialize = TRUE,
  file = FALSE,
  length = Inf,
  skip = "auto",
  ascii = FALSE,
  raw = FALSE,
  seed = 0,
  errormode = c("stop", "warn", "silent"),
  serializeVersion = .getSerializeVersion()
)
}
\arguments{
\item{object}{An R object to calculate a hash value for}

\item{algo}{The hashing algorithm to be used. See digest::digest for options}

\item{serialize}{Whether to serialize the object first. Default TRUE}

\item{file}{A filename to read from. Default FALSE}

\item{length}{The number of characters to read from file. Default NULL}

\item{skip}{Number of bytes to skip before reading file. Default 0}

\item{ascii}{If TRUE, return hash in ASCII format. Default FALSE}

\item{raw}{If TRUE, return raw vector. Default FALSE}

\item{seed}{Random seed for hash. Default 0}

\item{errormode}{Error mode for file reading. Default "stop"}

\item{serializeVersion}{Version used in serialization. Default NULL}
}
\value{
A character vector of hash values
}
\description{
This function applies the digest::digest function to each element of a
vector, returning a vector of hash values.
}
\examples{
vdigest(c("a", "b", "c"))
vdigest(as.character(iris$Species))
}
