% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency_check.R
\name{consistency_check}
\alias{consistency_check}
\title{Consistency Check Function}
\usage{
consistency_check(
  data,
  tests,
  cases,
  save_plot = FALSE,
  plot_path = NULL,
  target_language = "en",
  source_language = "en",
  lang_cache_path = tempdir(),
  compress_image = TRUE,
  image_overwrite = TRUE,
  compression_speed = 1,
  compression_verbose = TRUE
)
}
\arguments{
\item{data}{A data frame containing the test and case data.}

\item{tests}{A character vector specifying the column names for the test
data.}

\item{cases}{A character vector specifying the column names for the case
data. The length of 'tests' and 'cases' must be the same, and each element
in 'tests' corresponds to an element in 'cases'.}

\item{save_plot}{Logical. If TRUE, the plot will be saved to disk. Default
is FALSE.}

\item{plot_path}{Character. Directory path where the plot should be saved.
Required if save_plot is TRUE. Default is NULL.}

\item{target_language}{A character string specifying the language for plot
labels. Defaults to "en" (English). Use ISO 639-1 language codes.}

\item{source_language}{Source language code, defaults to "en"}

\item{lang_cache_path}{Path for translation cache, defaults to tempdir()}

\item{compress_image}{Logical. If TRUE, the saved image will be compressed.
Default is TRUE.}

\item{image_overwrite}{Logical. If TRUE, existing image will be overwritten.
Default is TRUE.}

\item{compression_speed}{Integer from 1-10. Speed of compression
(1=slow/high compression). Default is 1.}

\item{compression_verbose}{Logical. If TRUE, compression details will be
printed. Default is TRUE.}
}
\value{
A \code{\link[=ggplot2]{ggplot2::ggplot()}} object showing
the consistency between the number of tests and cases. The x-axis
represents the cases, and the y-axis represents the tests. Each facet
represents a disease, and the diagonal line shows where the number of
tests equals the number of cases.
}
\description{
This function performs a consistency check to ensure that the number of tests
is greater than the number of cases for given columns in a dataset. It
returns a \code{\link[=ggplot2]{ggplot2}} object visualizing the results.
}
\examples{
# check the consistency between malaria tests and cases

# get path
path <- system.file(
  "extdata",
  "fake_epi_df_togo.rds",
  package = "epiCleanr"
)

# get example data
fake_epi_df_togo <- read(path)

consistency_check(
  fake_epi_df_togo,
  tests = c("malaria_tests", "cholera_tests"),
  cases = c("malaria_cases", "cholera_cases")
)

# Save the plot
consistency_check(
  fake_epi_df_togo,
  tests = c("malaria_tests", "cholera_tests"),
  cases = c("malaria_cases", "cholera_cases"),
  save_plot = TRUE,
  plot_path = tempdir()
)

}
