% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_rate_plot.R
\name{get_translated_terms}
\alias{get_translated_terms}
\title{Get translated terms for plot filenames}
\usage{
get_translated_terms(
  target_language,
  source_language,
  lang_cache_path,
  x_var,
  vars_of_interest,
  save_title_prefix,
  data
)
}
\arguments{
\item{target_language}{Target language code}

\item{source_language}{Source language code}

\item{lang_cache_path}{Path for translation cache}

\item{x_var}{X-axis variable}

\item{vars_of_interest}{Variables being visualized}

\item{save_title_prefix}{A string prefix for the plot title and filename.
If NULL, a default prefix will be used based on the visualization type.}

\item{data}{Original data for year range}
}
\value{
List of translated terms
}
\description{
Get translated terms for plot filenames
}
